Review report + corrections

+++ src/main/java/note/repository/NoteRepositoryMock.java
@@ -40,7 +40,7 @@ public class NoteRepositoryMock implements NoteRepository{
 			throw new ClasaException(Constants.invalidNota);
 		if(nota.getNota() != (int)nota.getNota())
 			throw new ClasaException(Constants.invalidNota);
-		if(nota.getNrmatricol() != (int)nota.getNrmatricol())
+		if(nota.getNrmatricol() != nota.getNrmatricol())
 			throw new ClasaException(Constants.invalidNrmatricol);
 		return true;
 	}
@@ -59,7 +59,7 @@ public class NoteRepositoryMock implements NoteRepository{
 			String line;
 			while ((line = br.readLine()) != null) {
 				String[] values = line.split(";");
-				Nota nota = new Nota(Double.parseDouble(values[0]), values[1], Double.parseDouble(values[2]));
+				Nota nota = new Nota(Integer.parseInt(values[0]), values[1], Double.parseDouble(values[2]));
 				note.add(nota);
 			}
 			br.close();
